bf_backend/requirements.txt
django==1.9
celery==3.1.17
django-celery==3.1.17
django-bootstrap3==7.1.0
django-reversion==2.0.0
django-crispy-forms==1.6.1
django-formtools==1.0
httplib2==0.9.2
xadmin==0.6.1
django-debug-toolbar==1.6
djangorestframework==3.5.1
django_filters==0.2.1
django-filter==0.15.3
django-extensions==1.7.5
django-datatable==0.2.1